reorganising filenames and structure
