+++ dataworkspace/dataworkspace/templates/datasets/subscriptions/step_1_start.html
@@ -1,0 +1,46 @@
+{% extends '_main.html' %}
+
+{% block breadcrumbs %}
+  <div class="govuk-grid-row">
+    <div class="govuk-grid-column-three-quarters">
+      <div class="govuk-breadcrumbs">
+        <ol class="govuk-breadcrumbs__list">
+          <li class="govuk-breadcrumbs__list-item">
+            <a class="govuk-breadcrumbs__link" href="/">Home</a>
+          </li>
+          <li class="govuk-breadcrumbs__list-item">
+            <a href="{{ dataset.get_absolute_url }}">
+              {{ dataset.name }}
+            </a>
+          </li>
+          <li class="govuk-breadcrumbs__list-item">
+            Subscribe
+          </li>
+        </ol>
+      </div>
+    </div>
+  </div>
+{% endblock %}
+
+{% block content %}
+  <div class="govuk-grid-row">
+    <div class="govuk-grid-column-two-thirds">
+      <h1 class="govuk-heading-xl">Get updated when this dataset changes
+      </h1>
+
+      <p class="govuk-body">
+        You’ll get emails when we add or update dataset about: <a href="{{ dataset.get_absolute_url }}">
+              {{ dataset.name }}
+            </a>
+        <br>
+
+      </p>
+
+      <div class="govuk-form-group">
+        <a class="govuk-button" href="{% url "datasets:subscription_options" dataset.id %}">Continue</a>
+      </div>
+    </div>
+
+
+  </div>
+{% endblock %}
