+++ dataworkspace/dataworkspace/templates/datasets/subscriptions/step_2_options.html
@@ -25,7 +25,6 @@
 {% block content %}
   <div class="govuk-grid-row">
     <div class="govuk-grid-column-two-thirds">
-      <span class="govuk-caption-xl">{{ dataset.name }}</span>
       <h1 class="govuk-heading-l">What changes would you like to get emails about?
       </h1>
