+++ dataworkspace/dataworkspace/templates/partials/subscription_link.html
@@ -1,9 +1,9 @@
 {% if subscription.current_user_is_subscribed %}
-  <a href="{% url 'datasets:subscription_options' dataset.id %}" class="govuk-link">
+  <a href="{% url 'datasets:subscription_start' dataset.id %}" class="govuk-link">
     {% include "partials/subscription_icon.html" with text="Change your email preferences or unsubscribe" %}
   </a>
 {% else %}
-  <a href="{% url 'datasets:subscription_options' dataset.id %}" class="govuk-link">
+  <a href="{% url 'datasets:subscription_start' dataset.id %}" class="govuk-link">
     {% include "partials/subscription_icon.html" with text="Get updated when this dataset changes" %}
   </a>
 {% endif %}
